"example.sfc" contains some constructs possible in this sfc-language. output.txt is the output of java slime.sfcparser.ParserTest example.sfc > output.txt. The last line was added by hand, you may use (if your os does support this) 2>&1 for outputting the last line into that file too. The error that occurs in an error within the slime.utils.PrettyPrint,
I will try to fix this error now, but the fixed version
will reside in slime/sfcparser, because i do not want to
change someones code without a prior discussion.
Btw. someone modified my SFC.lex with the result that
the whilestatement did not function for a long period of
time - thank you very much.

My advice for the next semester would be to use LOGIN-
accounts for cvs. In my humble opinion it was Thomas
Richter who was the first to propose this. It makes the
logs more useful and avoids problems with $CVS_RSH.
(mwe)